requirements.txt

# Core AI
openai>=1.12.0

# Gmail + Google APIs
google-api-python-client>=2.125.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# FastAPI (for building API endpoints later)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Utilities
python-dotenv>=1.0.1   # for managing .env secrets
requests>=2.31.0       # for HTTP requests if needed
